% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\docType{data}
\name{members}
\alias{members}
\title{Members of The Travel Reports}
\format{Data frame with columns
\describe{
\item{id}{Report id, join this id with table 'reports'}
\item{member_name}{}
\item{member_office}{}
\item{member_unit}{}
\item{member_job}{}
\item{member_level}{}
}}
\usage{
members
}
\description{
Extracted from reports
}
\examples{
members
}
\keyword{datasets}

