% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\docType{data}
\name{reports}
\alias{reports}
\title{ROC Government Travel Reports}
\format{Data frame with columns
\describe{
 \item{gov}
 \item{id}{Unique identifier of a report}
 ...
}}
\source{
\url{https://drive.google.com/drive/folders/0B44pNgi6s_pGRDlsSWdUck5fZ1k}, downloaded 2016-08-22
}
\usage{
reports
}
\description{
This data set covers reports of government travel.
See \url{https://g0v.hackpad.com/E0G6gZDQ2ZZ} for more detail.
}
\examples{
  reports_df<- reports
}
\keyword{datasets}

